+++ config/config.js
@@ -3,7 +3,7 @@ const os = require('os')
 
 const DEFAULT = {
   PROTOCOL: 'http',
-  DOMAIN: 'localhost',
+  DOMAIN: '127.0.0.1',
   PORT: 5000,
   CACHE: path.join(os.homedir(), 'mbtiles'),
   VERBOSE: false